+++ src/app/views/amap-polyline-demo/methods/code.ts
@@ -1,6 +1,6 @@
+export const TS = `\
 import { Component, OnInit, ViewChild } from '@angular/core';
 import { AmapPolylineDirective } from 'ngx-amap';
-import { TS, HTML } from './code';
 
 @Component({
   selector: 'app-methods',
@@ -8,9 +8,6 @@ import { TS, HTML } from './code';
   styleUrls: ['./methods.component.scss']
 })
 export class MethodsComponent implements OnInit {
-  code_ts = TS;
-  code_html = HTML;
-
   lineArr = [
     [116.368904, 39.913423],
     [116.382122, 39.901176],
@@ -33,4 +30,20 @@ export class MethodsComponent implements OnInit {
       polyline.getExtData().then(v => console.log('getExtData():', v));
     }
   }
-}
+}`;
+
+export const HTML = `\
+<button type="button" class="btn btn-outline-primary" (click)="getMethods(myPolyline)">调用Getter方法</button>
+<button type="button" class="btn btn-outline-primary" (click)="editor = !editor">打开/关闭编辑功能</button>
+<hr>
+<ngx-amap class="demo-map" [center]="[116.397428, 39.90923]" [zoom]="13">
+  <amap-polyline #myPolyline="polyline"
+    [editor]="editor"
+    [path]="lineArr"
+    [strokeColor]="'#3366FF'"
+    [strokeOpacity]="1"
+    [strokeWeight]="5"
+    [strokeStyle]="'solid'"
+    [strokeDasharray]="[10, 5]">
+  </amap-polyline>
+</ngx-amap>`;